refactor: migrate interface/contrib to coins dirs
