+++ basicswap/interface/bch/contrib/cashaddress.py
