+++ basicswap/interface/bch/contrib/script.py
