+++ basicswap/interface/firo/contrib/siphash.py
