+++ basicswap/interface/nav/contrib/mininode.py
