+++ basicswap/interface/nav/contrib/siphash.py
