+++ basicswap/interface/pivx/contrib/siphash.py
